apps/backend/pyproject.toml
[project]
name = "aeronlp-backend"
version = "0.1.0"
description = "ZhiJian-AeroNLP backend service"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "pydantic[email]>=2.0.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
